gateway/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
asyncpg==0.30.0
pyjwt[crypto]==2.10.1
argon2-cffi==23.1.0
httpx==0.28.1
pydantic-settings==2.7.1
python-multipart==0.0.20
websockets==14.2
